Support python 2 again
