+++ test/tests.py
@@ -1,11 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import nose.tools as nt
-try:
-    from asyncio import get_event_loop
-    from asyncio import sleep as async_sleep
-except ImportError:
-    pass
 
 from pyee import EventEmitter
 
@@ -37,38 +32,6 @@ def test_emit():
     with nt.assert_raises(ItWorkedException) as it_worked:
         ee.emit('event', 'emitter is emitted!', error=True)
 
-def test_async_emit():
-    if not get_event_loop:
-        return
-
-    """Test that event_emitters can handle wrapping coroutines
-    """
-    ee = EventEmitter()
-    loop = get_event_loop()
-
-    class SenseWasCalled():
-        def __init__(self):
-            self.was_called = False
-
-        def am_calling(self):
-            self.was_called = True
-
-        def assert_was_called(self):
-            nt.assert_true(self.was_called)
-
-    sensor = SenseWasCalled()
-
-
-    @ee.on('event')
-    async def event_handler():
-        sensor.am_calling()
-
-
-    ee.emit('event')
-    loop.run_until_complete(async_sleep(1))
-
-    sensor.assert_was_called()
-    
 
 def test_emit_error():
     ee = EventEmitter()
